+++ library/feature/about/src/main/kotlin/com/mihaicristiancondrea/android/libs/apptoolkit/feature/about/di/AboutModule.kt
@@ -81,7 +81,7 @@ val aboutModule: Module = module {
     viewModel {
         AboutViewModel(
             aboutRepository = get(),
-            copyDeviceInfo = get(),
+            copyDeviceInfoUseCase = get(),
             dispatchers = get(),
             firebaseController = get(),
         )